Quiz.app/requirements.txt
flask==3.1.0
groq==0.10.0
python-dotenv==1.0.1
